taskmaster/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
pydantic[email]>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.7
slowapi>=0.1.9
websockets>=12.0
httpx>=0.26.0
pytest>=7.4.4
pytest-asyncio>=0.23.3
aiosqlite>=0.19.0
greenlet>=3.0.3
